# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SpaceArt API",
        "version": "2.0.0"
    },
    "servers": [
        {
            "url": "http://api-spaceart.local/api/"
        }
    ],
    "paths": {
...
        "/selective/{selective}/candidate": {
            "post": {
                "tags": [
                    "Candidate"
                ],
                "summary": "Store candidature",
                "description": "Strore candidature on selective",
                "operationId": "9c54899a0aaace5941f343c13d631f6c",
                "parameters": [
                    {
                        "name": "selective",
                        "in": "path",
                        "description": "Selective id",
                        "style": "form",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/SelectiveCandidateStore"
                },
                "responses": {
                    "201": {
                        "description": "Candidature finished successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "default": "Candidature was created"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/SelectiveCandidate"
                                            }
                                        },
                                        "fails": {
                                            "type": "boolean",
                                            "default": "false"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "$ref": "#/components/securitySchemes/Sanctum"
                    }
                ]
            }
        },
        "/selective": {
            "get": {
                "tags": [
                    "Selective"
                ],
                "summary": "List selectives",
                "description": "Fetch selectives on database",
                "operationId": "b842f2148613abfd36ecb03f5aa4e118",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/Offset"
                    },
                    {
                        "$ref": "#/components/parameters/Limit"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operation was realized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Selective"
                                            }
                                        },
                                        "fails": {
                                            "type": "boolean",
                                            "default": "false"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Selective"
                ],
                "summary": "Store selective",
                "description": "Store selective on database",
                "operationId": "8df69ff976dccbce85ee35ddf0fbf789",
                "requestBody": {
                    "$ref": "#/components/requestBodies/SelectiveStore"
                },
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/ReturnSelective"
                    }
                },
                "security": [
                    {
                        "$ref": "#/components/securitySchemes/Sanctum"
                    }
                ]
            }
        },
        "/selective/{id}": {
            "get": {
                "tags": [
                    "Selective"
                ],
                "summary": "Fetch unique selective",
                "description": "Fetch selective on database",
                "operationId": "f8fc3862f421565dfd8e7dcb28482b2b",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/Id"
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/ReturnSelective"
                    }
                },
                "security": [
                    {
                        "$ref": "#/components/securitySchemes/Sanctum"
                    }
                ]
            },
            "put": {
                "tags": [
                    "Selective"
                ],
                "summary": "Update selective",
                "description": "Update selective on database",
                "operationId": "802cf399957a7b0184bbb3de019e0173",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/Id"
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/SelectiveUpdate"
                },
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/ReturnSelective"
                    }
                },
                "security": [
                    {
                        "$ref": "#/components/securitySchemes/Sanctum"
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Selective"
                ],
                "summary": "Delete selective",
                "description": "Delete selective on database",
                "operationId": "f66a228635f8ce532147a41f96d116f7",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/Id"
                    }
                ],
                "responses": {
                    "204": {
                        "$ref": "#/components/responses/204"
                    }
                },
                "security": [
                    {
                        "$ref": "#/components/securitySchemes/Sanctum"
                    }
                ]
            }
        },
        "/selective/{id}/update": {
            "post": {
                "tags": [
                    "Selective"
                ],
                "summary": "[PUT] /selective/{id} alias ",
                "description": "Redirect to [PUT] /selective/{id}",
                "operationId": "8cdcac6947cf4f0b5968f15e0c8ced07",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/Id"
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Redirected to [PUT] /selective/{id}"
                    }
                },
                "security": [
                    {
                        "$ref": "#/components/securitySchemes/Sanctum"
                    }
                ]
            }
        },
        "/selective/{id}/delete": {
            "post": {
                "tags": [
                    "Selective"
                ],
                "summary": "[DELETE] /selective/{id} alias ",
                "description": "Redirect to [DELETE] /selective/{id}",
                "operationId": "7f8749bacabbcfe0f94d92111dd5d405",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/Id"
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Redirected to [DELETE] /selective/{id}"
                    }
                },
                "security": [
                    {
                        "$ref": "#/components/securitySchemes/Sanctum"
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Art": {
                "description": "Schema of Model Art",
                "properties": {
                    "name": {
                        "type": "enum",
                        "enum": [
                            "acting",
                            "dance",
                            "music",
                            "painting",
                            "sculpture"
                        ]
                    },
                    "artists_count": {
                        "description": "number of relationed artists",
                        "type": "integer"
                    },
                    "agreements_count": {
                        "description": "number of relationed agreements",
                        "type": "integer"
                    },
                    "selectives_count": {
                        "description": "number of relationed selectives",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "Artist": {
                "description": "Schema of Model Artist User",
                "properties": {
                    "cpf": {
                        "type": "string",
                        "example": "40033796599"
                    },
                    "bithday": {
                        "type": "string",
                        "format": "date",
                        "example": "01/01/1970"
                    },
                    "wage": {
                        "type": "number",
                        "example": "100"
                    },
                    "art": {
                        "$ref": "#/components/schemas/Art"
                    }
                },
                "type": "object"
            },
            "Enterprise": {
                "description": "Schema of Model Enterprise User",
                "properties": {
                    "cnpj": {
                        "type": "string",
                        "example": "01499146000196"
                    },
                    "company_name": {
                        "type": "string",
                        "example": "José e Gabriela Esportes ME"
                    },
                    "agreements": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Agreement"
                        }
                    },
                    "selectives": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Selective"
                        }
                    }
                },
                "type": "object"
            },
...
            "Selective": {
                "description": "Schema of Model Selective",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "enterprise": {
                        "$ref": "#/components/schemas/Enterprise"
                    },
                    "art": {
                        "$ref": "#/components/schemas/Art"
                    },
                    "title": {
                        "description": "Comercial name of selective",
                        "type": "string",
                        "example": "Summer 2024's selective"
                    },
                    "note": {
                        "description": "Description of service",
                        "type": "string",
                        "example": "Rock musical show for 2 hours"
                    },
                    "start_moment": {
                        "description": "Moment of open selective",
                        "type": "string",
                        "format": "date",
                        "example": "01/01/2025 00:00"
                    },
                    "end_moment": {
                        "description": "Moment of open selective",
                        "type": "string",
                        "format": "date",
                        "example": "31/12/2025 23:59"
                    },
                    "price": {
                        "description": "Value of fee",
                        "type": "number",
                        "format": "float",
                        "example": "1000.00"
                    },
                    "candidates": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/SelectiveCandidate"
                        }
                    }
                },
                "type": "object"
            },
            "SelectiveCandidate": {
                "description": "Schema of Selective Candidate",
                "properties": {
                    "artist": {
                        "$ref": "#/components/schemas/User"
                    },
                    "selective": {
                        "$ref": "#/components/schemas/Selective"
                    }
                },
                "type": "object"
            },
            "User": {
                "description": "Schema of Model User",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "name": {
                        "description": "Username",
                        "type": "string",
                        "example": "José Carlos"
                    },
                    "active": {
                        "description": "Access to account is enabled",
                        "type": "boolean",
                        "default": "true"
                    },
                    "email": {
                        "description": "Access email address",
                        "type": "string",
                        "example": "example@org.net"
                    },
                    "phone": {
                        "description": "User's mobile number",
                        "type": "string",
                        "example": "71988469787"
                    },
                    "type": {
                        "description": "Account type",
                        "type": "enum",
                        "enum": [
                            "artist",
                            "enterprise"
                        ]
                    },
                    "postal_code": {
                        "description": "Brazilian zip code",
                        "type": "string",
                        "example": "41000000"
                    },
                    "state": {
                        "description": "Brazilian state acronym",
                        "type": "string",
                        "example": "BA"
                    },
                    "city": {
                        "type": "string",
                        "example": "Salvador"
                    },
                    "neighborhood": {
                        "type": "string",
                        "example": "Piatã"
                    },
                    "street": {
                        "type": "string",
                        "example": "Av. Orlando Gomes"
                    },
                    "address_complement": {
                        "description": "Additional  information of address",
                        "type": "string",
                        "example": "Beside of SESI Saúde"
                    },
                    "image": {
                        "description": "Image in base64 or URL",
                        "type": "string"
                    },
                    "slug": {
                        "description": "User's website address",
                        "type": "string",
                        "example": "https://spaceart-lemon.vercel.app/user/2",
                        "nullable": true
                    },
                    "biography": {
                        "description": "Short presentation text",
                        "type": "string",
                        "nullable": true
                    },
                    "received_rates_avg_score": {
                        "type": "number",
                        "nullable": true
                    },
                    "artist_account_data": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/Artist"
                            }
                        ],
                        "nullable": true
                    },
                    "enterprise_account_data": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/Enterprise"
                            }
                        ],
                        "nullable": true
                    },
                    "sent_rates": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Rate"
                        }
                    },
                    "received_rates": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Rate"
                        }
                    },
                    "posts": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Post"
                        }
                    }
                },
                "type": "object"
            }
        },
        "responses": {
...
            "204": {
                "description": "Resource was disabled",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string"
                                },
                                "fails": {
                                    "type": "boolean",
                                    "default": "false"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
            "ReturnSelective": {
                "description": "Operation was realized",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string"
                                },
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Selective"
                                    },
                                    "maxItems": 1
                                },
                                "fails": {
                                    "type": "boolean",
                                    "default": "false"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
        },
        "parameters": {
            "Limit": {
                "name": "limit",
                "in": "query",
                "description": "Limit per page",
                "style": "form",
                "schema": {
                    "type": "integer",
                    "nullable": true
                }
            },
            "Offset": {
                "name": "offset",
                "in": "query",
                "description": "Offset for search",
                "style": "form",
                "schema": {
                    "type": "integer",
                    "nullable": true
                }
            },
            "Id": {
                "name": "id",
                "in": "path",
                "description": "Resource id",
                "style": "form",
                "schema": {
                    "type": "integer"
                }
            },
...
        },
        "requestBodies": {
...
            "SelectiveCandidateStore": {
                "content": {
                    "application/json": {
                        "schema": {
                            "required": [
                                "artist_id"
                            ],
                            "properties": {
                                "artist_id": {
                                    "type": "integer"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "SelectiveStore": {
                "content": {
                    "application/json": {
                        "schema": {
                            "required": [
                                "enterprise_id",
                                "title",
                                "start_moment",
                                "end_moment",
                                "art",
                                "note",
                                "price"
                            ],
                            "properties": {
                                "enterprise_id": {
                                    "type": "integer"
                                },
                                "title": {
                                    "type": "string",
                                    "example": "Summer 2025"
                                },
                                "start_moment": {
                                    "type": "string",
                                    "format": "d/m/Y H:i",
                                    "example": ""
                                },
                                "end_moment": {
                                    "type": "string",
                                    "format": "d/m/Y H:i",
                                    "example": ""
                                },
                                "art": {
                                    "type": "enum",
                                    "enum": [
                                        "acting",
                                        "dance",
                                        "music",
                                        "painting",
                                        "sculpture"
                                    ]
                                },
                                "note": {
                                    "type": "string",
                                    "example": "perfect!!"
                                },
                                "price": {
                                    "type": "number"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "SelectiveUpdate": {
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "title": {
                                    "type": "string",
                                    "example": "Summer 2025"
                                },
                                "start_moment": {
                                    "type": "string",
                                    "format": "d/m/Y H:i",
                                    "example": ""
                                },
                                "end_moment": {
                                    "type": "string",
                                    "format": "d/m/Y H:i",
                                    "example": ""
                                },
                                "note": {
                                    "type": "string",
                                    "example": "perfect!!"
                                },
                                "price": {
                                    "type": "number"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
        },
        "securitySchemes": {
            "Sanctum": {
                "type": "apiKey",
                "scheme": {
                    "bearerToken": []
                }
            }
        }
    }
}